@@ -10,28 +10,36 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.0.0",
+    "@angular/animations": "^17.1.1",
     "@angular/cdk": "^17.0.1",
-    "@angular/common": "^17.0.0",
-    "@angular/compiler": "^17.0.0",
-    "@angular/core": "^17.0.0",
-    "@angular/forms": "^17.0.0",
-    "@angular/platform-browser": "^17.0.0",
-    "@angular/platform-browser-dynamic": "^17.0.0",
-    "@angular/router": "^17.0.0",
-    "@angular/service-worker": "^17.0.0",
+    "@angular/common": "^17.1.1",
+    "@angular/compiler": "^17.1.1",
+    "@angular/core": "^17.1.1",
+    "@angular/forms": "^17.1.1",
+    "@angular/platform-browser": "^17.1.1",
+    "@angular/platform-browser-dynamic": "^17.1.1",
+    "@angular/router": "^17.1.1",
+    "@angular/service-worker": "^17.1.1",
     "@ctrl/ngx-emoji-mart": "^9.2.0",
-    "@ngrx/component": "^17.0.0",
-    "@ngrx/effects": "^17.0.0",
-    "@ngrx/entity": "^17.0.0",
-    "@ngrx/eslint-plugin": "^17.0.0",
-    "@ngrx/router-store": "^17.0.0",
-    "@ngrx/signals": "^17.0.0",
-    "@ngrx/store": "^17.0.0",
-    "@ngrx/store-devtools": "^17.0.0",
+    "@ngrx/component": "^17.1.0",
+    "@ngrx/effects": "^17.1.0",
+    "@ngrx/entity": "^17.1.0",
+    "@ngrx/eslint-plugin": "^17.1.0",
+    "@ngrx/router-store": "^17.1.0",
+    "@ngrx/signals": "^17.1.0",
+    "@ngrx/store": "^17.1.0",
+    "@ngrx/store-devtools": "^17.1.0",
+    "@tailwindcss/typography": "^0.5.10",
+    "clipboard": "^2.0.11",
+    "emoji-toolkit": "^8.0.0",
+    "katex": "^0.16.9",
+    "marked": "^9.1.6",
     "material-icons": "^1.13.12",
+    "mermaid": "^10.6.1",
     "ngx-cookie-service": "^17.0.0",
     "ngx-image-cropper": "^7.1.2",
+    "ngx-markdown": "^17.1.1",
+    "prismjs": "^1.29.0",
     "rxjs": "~7.8.0",
     "socket.io-client": "^4.7.2",
     "three": "^0.158.0",
@@ -39,10 +47,10 @@
     "zone.js": "~0.14.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.0.3",
-    "@angular/cli": "^17.0.3",
-    "@angular/compiler-cli": "^17.0.0",
-    "@ngrx/schematics": "^17.0.0",
+    "@angular-devkit/build-angular": "^17.1.1",
+    "@angular/cli": "^17.1.1",
+    "@angular/compiler-cli": "^17.1.1",
+    "@ngrx/schematics": "^17.1.0",
     "@types/jasmine": "~5.1.0",
     "@types/socket.io-client": "^3.0.0",
     "@types/three": "^0.158.3",
