@@ -1,4 +1,4 @@
-import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
+import { ChangeDetectionStrategy, Component, input } from '@angular/core';
 import CardComponent from '../../../../../../shared/components/card/card.component';
 import ImgAvatarComponent from '../../../../../../shared/components/img-avatar/img-avatar.component';
 import { ContactDetails } from '../../../../../../stores/contacts/contacts.model';
@@ -16,10 +16,10 @@ import { User } from '../../../../../../stores/user/user.model';
       <ng-container ngProjectAs="body">
         <div class="w-full flex justify-between items-center">
           <div class="flex gap-4">
-            <avatar class="pointer" [url]="contact?.profile?.url" />
+            <avatar class="pointer" [url]="contact().profile.url" />
             <section class="flex flex-col gap-1">
-              <div class="text-xs">{{ contact?.name }}</div>
-              <div class="text-xs">{{ '@' + contact?.username }}</div>
+              <div class="text-xs">{{ contact().name }}</div>
+              <div class="text-xs">{{ '@' + contact().username }}</div>
             </section>
           </div>
           <div class="flex gap-4">
@@ -33,6 +33,6 @@ import { User } from '../../../../../../stores/user/user.model';
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 class ContactCardComponent {
-  @Input() contact!: ContactDetails | User;
+ contact = input.required< ContactDetails | User>();
 }
 export default ContactCardComponent;
