@@ -1,4 +1,4 @@
-import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
+import { ChangeDetectionStrategy, Component, input } from '@angular/core';
 import CardComponent from '../../../../../../../../shared/components/card/card.component';
 import ImgAvatarComponent from '../../../../../../../../shared/components/img-avatar/img-avatar.component';
 import { User } from '../../../../../../../../stores/user/user.model';
@@ -13,15 +13,15 @@ import { NgIf } from '@angular/common';
       <ng-container ngProjectAs="body">
         <div class="w-full flex justify-between items-center">
           <div class="flex gap-4">
-            <avatar class="pointer" [url]="user?.profile?.url ?? ''" />
+            <avatar class="pointer" [url]="user().profile?.url ?? ''" />
             <section class="flex flex-col gap-1">
-              @if(user?.name) {
-              <div class="text-xs">{{ user?.name }}</div>
+              @if(user().name) {
+              <div class="text-xs">{{ user().name }}</div>
               } @else {
               <div class="text-xs w-52 h-4 animate:pulse rounded bg-slate-400"></div>
               }
-              @if(user?.name) {
-                <div class="text-xs">{{ '@' + user?.username }}</div>
+              @if(user().name) {
+                <div class="text-xs">{{ '@' + user().username }}</div>
               } @else {
                 <div class="text-xs w-36 h-4 animate:pulse rounded bg-slate-400"></div>
               }
@@ -38,6 +38,6 @@ import { NgIf } from '@angular/common';
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 class UserCardComponent {
-  @Input() user: User | undefined;
+ user = input<User>();
 }
 export default UserCardComponent;
