@@ -1,7 +1,7 @@
 import {
   ChangeDetectionStrategy,
   Component,
-  Input,
+  input,
   Signal,
   inject,
   signal,
@@ -23,7 +23,7 @@ import ImageComponent from '../image/image.component';
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 class NavbarComponent {
-  @Input() isUser = false;
+  isUser = input<boolean>(false);
   readonly Routes: Signal<RoutesInterface> = signal(ClientRoutes);
   readonly AppIcon: Signal<string> = signal(App.ICON);
   readonly mode = inject(ModeService);
